splitting energySupplyComponents from newComponents.. Energy Supply models are not very useful for tespy I believe
(cherry picked from commit bed22e1da7fdaacc77462c70bd004102c4c3b5b1)
