+++ intermediateFoodTests/energySupply/energy-supply_energy-merge3.py
@@ -6,7 +6,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from tespy.components import Separator,Merge,CycleCloser,Valve,Splitter
-from tespy.components.newcomponents import *
+from tespy.components.energySupplyComponents import *
 
 import logging
 logging.basicConfig(level=logging.DEBUG)
