+++ src/tespy/components/__init__.py
@@ -30,3 +30,4 @@ from .turbomachinery.turbine import Turbine  # noqa: F401
 
 # New components
 from .newcomponents import *
+from .energySupplyComponents import *
